Convert tests for FailTransfer
